worker/requirements.txt
pymongo==4.8.0
redis==5.0.8
python-dotenv==1.0.1
